% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/task_functions.R
\name{finaltask_combine_and_clean}
\alias{finaltask_combine_and_clean}
\title{Final Revision: Combine and Clean Single-Cell Metadata}
\usage{
finaltask_combine_and_clean(dt_clusters, dt_celltypes)
}
\arguments{
\item{dt_clusters}{A data.table containing 'cell' ID and 'integration_cluster' (from SeuratIntegration file).}

\item{dt_celltypes}{A data.table containing 'cell' ID, 'cell_type', and 'sample_type' (N/T).}
}
\value{
A combined data.table ('master table') containing 'cell', 'integration_cluster', 'cell_type', and 'sample_type'.
}
\description{
Performs necessary cleaning on the 'cell' IDs (trimming, removing suffixes like _N/_T)
and executes an inner join between the clusters data and the cell types data to create a master table.
}
